Готовое решение после приведения порядка в коде и переноса переиспользуемых методов получения счетов и клиентов в хелпер

+++ Bank.Application/Clients/Commands/DeleteClient/DeleteClientCommandHandler.cs
@@ -17,7 +17,10 @@ public class DeleteClientCommandHandler : IRequestHandler<DeleteClientCommand>
     {
         
         var client = await _context.Clients.FirstOrDefaultAsync(c => c.Id == request.Id, cancellationToken);
-        _context.Clients.Remove(client);
+        if (client != null)
+        {
+            _context.Clients.Remove(client);
+        }
         
         await _context.SaveChangesAsync(cancellationToken);
     }
